requirements.txt
asgiref==3.2.3
astroid==2.3.3
colorama==0.4.3
Django==3.0.2
django-choices==1.7.1
django-crispy-forms==1.8.1
django-mathfilters==0.4.0
django-model-utils==4.0.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
mysqlclient==1.4.6
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
rope==0.16.0
six==1.14.0
sqlparse==0.3.0
wrapt==1.11.2
